refactor: moving of types, renames and error updates
